requirements.txt
pandas
numpy
faker
pyarrow
pyspark
sqlalchemy
pyodbc
azure-identity
azure-storage-blob
jupyter
pytest
